Update README.md
The following files are for a python chatbot that can be trained with WhatsApp training data. The bot uses a SQL database to store response and remembers them with each new initialization of the bot. The code requires a virtual environment to be used in order to install the chatterbot framework.